# Conflicts:
#	.gitignore
#	chess/0-chess-moves/chess-board.png
#	chess/0-chess-moves/diagrams.uml
#	chess/1-chess-game/chess-board.png
#	chess/1-chess-game/diagrams.uml
#	chess/1-chess-game/getting-started.md
#	chess/3-web-api/getting-started.md
#	chess/4-database/getting-started.md
#	chess/Info Files/Phase 1/chess-board.png
#	chess/Info Files/Phase 1/class-structure.png
#	chess/Info Files/Phase 1/create-chess-project.png
#	chess/Info Files/Phase 1/diagrams.uml
#	chess/Info Files/Phase 1/directory.png
#	chess/Info Files/Phase 1/test-run-button.png
#	chess/Info Files/Phase 1/test-run-menu.png
#	chess/Info Files/Phase 2/chess-classes.png
#	chess/Info Files/Phase 2/example.uml
#	chess/Info Files/Phase 2/register-sequence-diagram.png
#	chess/Info Files/Phase 2/sequence-diagram-org.gif
#	chess/Info Files/Phase 2/server-class-structure.png
#	chess/Info Files/Phase 2/server-design-architecture.png
#	chess/Info Files/Phase 2/server-design.md
#	chess/Info Files/Phase 2/top-level.png
#	chess/Info Files/Phase 3/chess-server-webpage.png
#	chess/Info Files/Phase 3/data-access-classes.png
#	chess/Info Files/Phase 3/diagrams.uml
#	chess/Info Files/Phase 3/login-interaction.png
#	chess/Info Files/Phase 3/mark-test-root.png
#	chess/Info Files/Phase 4/database.md
#	chess/Info Files/Phase 4/existingDependencies.png
#	chess/Info Files/Phase 4/install-dependency.gif
#	chess/Info Files/Phase 5/ChessBoard.png
#	chess/Info Files/Phase 5/blackBoard.png
#	chess/Info Files/Phase 5/whiteBoard.png
#	chess/Info Files/Phase 6/Rodham-ChessWebSocketDesign.png
#	chess/Info Files/Phase 6/getting-started.md
#	chess/chess.md
#	chess/client/lib/tyrus-standalone-client-1.15.jar
#	chess/code-quality-rubric.md
#	chess/server/lib/passoff-dependencies.jar
#	chess/web/favicon.ico
#	chess/web/index.html
#	instruction/copying-objects/copying-objects.md
#	spelling-corrector/notsobig.txt
#	spelling-corrector/out/production/spelling-corrector/jars/apiguardian-api-1.1.0 2.jar
#	spelling-corrector/out/production/spelling-corrector/jars/junit-jupiter-api-5.6.2 2.jar
#	spelling-corrector/out/production/spelling-corrector/jars/junit-platform-commons-1.6.2 2.jar
#	spelling-corrector/out/production/spelling-corrector/jars/opentest4j-1.2.0 2.jar
#	spelling-corrector/src/passoff/SpellTest.java
#	spelling-corrector/src/passoff/TrieTest.java
#	spelling-corrector/src/spell/INode.java
#	spelling-corrector/src/spell/ISpellCorrector.java
#	spelling-corrector/src/spell/ITrie.java
#	spelling-corrector/word.txt
#	spelling-corrector/words.txt

+++ petshop/server/src/main/service/PetService.java
@@ -1,8 +1,9 @@
 package service;
 
 import dataaccess.DataAccess;
-import model.Pet;
+import model.*;
 import exception.ResponseException;
+import model.PetType;
 
 import java.util.Collection;
 
@@ -14,23 +15,40 @@ public class PetService {
         this.dataAccess = dataAccess;
     }
 
+    // Pet Shop is very simple.
+    // A more complicated application would do the business logic in the service.
+
     public Pet addPet(Pet pet) throws ResponseException {
+        if (pet.type() == PetType.DOG && pet.name().equals("fleas")) {
+            throw new ResponseException(ResponseException.Code.ClientError, "Error: no dogs with fleas");
+        }
         return dataAccess.addPet(pet);
     }
 
-    public Collection<Pet> listPets() throws ResponseException {
+    public PetList listPets() throws ResponseException {
         return dataAccess.listPets();
     }
 
     public Pet getPet(int id) throws ResponseException {
+        validateId(id);
         return dataAccess.getPet(id);
     }
 
     public void deletePet(Integer id) throws ResponseException {
+        validateId(id);
         dataAccess.deletePet(id);
     }
 
     public void deleteAllPets() throws ResponseException {
-        dataAccess.deleteAllPets();
+        Collection<Pet> pets = dataAccess.listPets();
+        if (!pets.isEmpty()) {
+            dataAccess.deleteAllPets();
+        }
+    }
+
+    private void validateId(int id) throws ResponseException {
+        if (id <= 0) {
+            throw new ResponseException(ResponseException.Code.ClientError, "Error: invalid pet ID");
+        }
     }
 }
